refactor: split semantika_cache.py into _cache_* sub-modules
semantika_cache.py reduced to thin re-export (34 lines).
New sub-modules:
  - _cache_db.py (98 lines): DB init, seeding, connection
  - _cache_lookup.py (172 lines): Three-layer lookup logic
  - _cache_api.py (82 lines): Wikidata API with circuit breaker
  - _cache_negative.py (44 lines): Negative cache tombstones
  - _cache_store.py (57 lines): Batch store + keyword gen
  - _cache_maintenance.py (72 lines): Invalidation, stats, repair
